logs_agent/requirements.txt
confluent-kafka==2.6.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
# Pin a genai version compatible with this code path
google-genai==1.56.0
